Replaced permission logic from activity to fragment
