+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/core/App.kt
@@ -37,8 +37,4 @@ class App :
 
     override fun onActivityDestroy() {
     }
-
-    fun getNavigationHolder() = cicerone.navigatorHolder
-
-    fun getRouter() = cicerone.router
 }
