+++ app/src/main/java/com/zhuravlevmikhail/a65appshomeproject/core/AppHostActivity.kt
@@ -1,16 +1,10 @@
 package com.zhuravlevmikhail.a65appshomeproject.core
 
-import android.content.pm.PackageManager
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.os.PersistableBundle
 import com.zhuravlevmikhail.a65appshomeproject.R
-import com.zhuravlevmikhail.a65appshomeproject.appManagers.ContactsManager
 import com.zhuravlevmikhail.a65appshomeproject.appManagers.PageManager
-import com.zhuravlevmikhail.a65appshomeproject.common.AppConst.PERMISSION_REQUEST_CODE_CONTACTS
-import ru.terrakok.cicerone.Navigator
 import ru.terrakok.cicerone.android.support.SupportAppNavigator
-import ru.terrakok.cicerone.commands.Command
 
 class AppHostActivity : AppCompatActivity(R.layout.activity_host) {
 
@@ -44,20 +38,6 @@ class AppHostActivity : AppCompatActivity(R.layout.activity_host) {
     override fun onDestroy() {
         pageManager.onDestroy()
         super.onDestroy()
-    }
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == PERMISSION_REQUEST_CODE_CONTACTS) {
-            if (grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                pageManager.onContactsAccessGranted()
-                App.instance.cicerone.router.newRootScreen(ContactsScreen())
-            }
-        }
     }
 
     private val navigator = SupportAppNavigator(this, R.id.fragmentsContainer)
